toggle menu and new table USER

+++ htdocs/xampp/inventory/removeItem.php
@@ -19,10 +19,10 @@ if ($conn->connect_error) {
 // Convert the query result into utf8
 $conn->query("SET character_set_results=utf8");
 
-$stmt = $conn->prepare("DELETE FROM inventory WHERE upc = ?");
+$stmt = $conn->prepare("DELETE FROM inventory WHERE upc = ? and user_id = ?");
 
 /* Binds variables to prepared statement */
-$stmt->bind_param('s', $upc);
+$stmt->bind_param('ss', $upc, $username);
 
 /* execute query */
 $stmt->execute() or die($stmt->error);
@@ -34,6 +34,6 @@ $stmt->free_result();
 $stmt->close();
 
 
-refreshInventory($conn);
+refreshInventory($conn,$username);
 
 $conn->close();
